Delete map step
