+++ liveHomeFrontend/pages/post/dashboard_user.js
@@ -11,7 +11,7 @@ import API from '../../src/utils/api';
 
 import styles from '../../src/styles/pages/post/dashboard_user.module.sass';
 
-const dashboardUser = ({properties}) => {
+const dashboardUser = ({ properties }) => {
   const inputSearch = setInputValue('');
 
   const handleSubmit = (event) => {
@@ -26,12 +26,12 @@ const dashboardUser = ({properties}) => {
           <Modal buttonText='Filter'>
             <p>Filter form</p>
           </Modal>
-          <SearchBar name='search' handleSubmit={handleSubmit} {...inputSearch}/>
+          <SearchBar name='search' handleSubmit={handleSubmit} {...inputSearch} />
         </div>
         <h1>Dashboard</h1>
         <div className={styles.offers}>
           {properties.map((item, key) => {
-            return <OffererCardPublication key={key} id={item.id} title={item.title} date={item.createdAt} />
+            return <OffererCardPublication key={key} id={item.id} title={item.title} date={item.createdAt} />;
           })}
         </div>
       </div>
@@ -39,13 +39,13 @@ const dashboardUser = ({properties}) => {
   );
 };
 
-export async function getStaticProps() {
+export async function getStaticProps () {
   const properties = await API.getProperties();
-  return { 
-    props: { 
-      properties: properties.data,
+  return {
+    props: {
+      properties: properties.data
     }
-  }
+  };
 }
 
 export default dashboardUser;
