--- liveHomeFrontend/src/styles/pages/post/new_post_step_four.module.sass
@@ -1,24 +1,0 @@
-@import "../../vars.sass"
-
-.container
-  margin-top: 30px
-  width: 100%
-  padding: 0 20px
-  box-sizing: border-box
-
-.buttons
-  display: grid
-  grid-template-columns: repeat(2, 1fr)
-  grid-gap: $spacing*5
-  width: 100%
-  margin-top: $spacing*2
-
-.info
-  color: $secondary-medium
-
-.thumbnails
-  width: 100%
-  display: flex
-  flex-direction: row
-  flex-wrap: wrap
-  justify-content: space-between
