+++ liveHomeFrontend/src/styles/pages/post/new_post_step_three.module.sass
@@ -12,3 +12,13 @@
   grid-gap: $spacing*5
   width: 100%
   margin-top: $spacing*2
+
+.info
+  color: $secondary-medium
+
+.thumbnails
+  width: 100%
+  display: flex
+  flex-direction: row
+  flex-wrap: wrap
+  justify-content: space-between
